FacebookSignedRequest.php
public abstract static class Application.Parameters
extends java.lang.Object

Encapsulates the set of parameters for an application. This includes arguments passed on the command line, unnamed parameters specified in a JNLP file, and <name,value> pairs specified in a JNLP file.

Note that the application and the preloader both get the same set of parameters for a given run of an application.